requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.34.0
packaging==24.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyYAML==6.0.1
QDarkStyle==3.2.3
QtPy==2.4.1
requests==2.32.3
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
webrtcvad==2.0.10
Werkzeug==3.0.3
